Models: Extracted LineFilterModel and PolygonFilterModel from feature_splitting_window.py
